enable render&sound error check on release update DX11 build

+++ src/Kernel/SHA1.cpp
@@ -28,7 +28,7 @@ namespace Mengine
             struct SHA1_CTX
             {
                 uint32_t state[5];
-                uint32_t count[2];
+                size_t count[2];
                 uint8_t buffer[64];
             };
             //////////////////////////////////////////////////////////////////////////
